+++ grow_your_roots/src/Components/User/User.js
@@ -12,15 +12,26 @@ const UserProfile = () => {
     Parse.User.current().authenticated = true;
   }
 
+
+  const goToProfile = () => {
+    Parse.User.current().authenticated = true;
+  }
+
+
   const goToBlog = () => {
     Parse.User.current().authenticated = true;
   }
+
     return (
         <section>
           <h1>Grow My Roots</h1>
           <div>
           <Link to="/">
             <button onClick={goToHome}>Home</button>
+          </Link>
+
+          <Link to="/users/profile">
+            <button onClick={goToProfile}>Profile</button>
           </Link>
           <Link to="/blog">
             <button onClick={goToBlog}>Blog</button>